feat: add hash to password
