+++ package.json
@@ -13,11 +13,13 @@
   "license": "ISC",
   "dependencies": {
     "@prisma/client": "6.6.0",
+    "bcrypt": "5.1.1",
     "fastify": "5.2.2"
   },
   "type": "module",
   "devDependencies": {
     "@eslint/js": "9.24.0",
+    "@types/bcrypt": "5.0.2",
     "@types/node": "22.14.0",
     "eslint": "9.24.0",
     "eslint-config-prettier": "10.1.1",
