+++ prisma/schema.prisma
@@ -12,6 +12,6 @@ model User {
   id          Int       @id @default(autoincrement())
   name        String    @db.VarChar(80)
   email       String    @unique @db.VarChar(100)
-  password    String    @db.VarChar(50)
+  password    String    @db.VarChar(255)
   birthDate   DateTime
 }
